+++ packages/components/src/components/sidebar/sidebar.tsx
@@ -13,15 +13,19 @@ const SIDEBAR_ITEM = '.sidebar__nav-item';
 export class Sidebar {
   @Element() el;
   @Prop() applicationName: string = ''
-  @Prop() termsOfUse: string = ""
-  @Prop() imprint: string = ""
   @Prop() initialCollapse: boolean = true
-  @Prop() privacyPolicy: string = ""
-  @Prop() target: string = "_self"
   @Prop() showFooter: boolean = true
+  @Prop() showHeader: boolean = true;
+  @Prop() termsOfUse: string = "#"
+  @Prop() imprint: string = "#"
+  @Prop() privacyPolicy: string = "#"
+  @Prop() target: string = "_blank"
+  @State() currentYear: number = new Date().getFullYear()
+  @Prop() copyrightText: string = '© 1999 - ' + this.currentYear + ' Infineon Technologies AG'
   @State() internalTermsofUse: string = ""
   @State() internalImprint: string = ""
   @State() internalPrivacyPolicy: string = ""
+
   @State() activeItem: HTMLElement | null = null;
 
   expandActiveItems(){
@@ -296,15 +300,17 @@ export class Sidebar {
 
 
   componentWillLoad() {
-    this.internalTermsofUse = this.termsOfUse || 'Not available';
-    this.internalPrivacyPolicy = this.privacyPolicy || 'Not available';
-    this.internalImprint = this.imprint || 'Not available';
+    this.internalTermsofUse = this.termsOfUse.trim();
+    this.internalPrivacyPolicy = this.privacyPolicy.trim();
+    this.internalImprint = this.imprint.trim();
   }
 
   render() {
     return (
       <div aria-label="a navigation sidebar" aria-value={this.applicationName} class='sidebar__container'>
         <div class='sidebar__top-container'>
+        {
+          this.showHeader && 
           <div class="sidebar__nav-bar">
             <div class="sidebar__nav-bar-logo">
               <div class='sidebar__nav-bar-logo-img'>
@@ -322,6 +328,7 @@ export class Sidebar {
               <div class='sidebar__nav-bar-logo-text'>{this.applicationName}</div>
             </div>
           </div>
+        }
           <div class="sidebar__nav-container">
             <slot />
           </div>
@@ -333,16 +340,26 @@ export class Sidebar {
             <div class="sidebar__footer-wrapper">
               <div class='sidebar__footer-wrapper-top-links'>
                 <div class="sidebar__footer-wrapper-top-line">
-                  <a target={this.target} href={this.internalTermsofUse}>Terms of use</a>
-                  <a target={this.target} href={this.internalImprint}>Imprint</a>
+                  {
+                    this.internalTermsofUse !== ''  && <a target={this.target} href={this.internalTermsofUse}>Terms of use</a>
+                  }
+                  {
+                    this.internalImprint !== '' && <a target={this.target} href={this.internalImprint}>Imprint</a>
+                  }
                 </div>
                 <div class="sidebar__footer-wrapper-bottom-line">
-                  <a target={this.target} href={this.internalPrivacyPolicy}>Privacy policy</a>
+                  {
+                    this.internalPrivacyPolicy !== '' && <a target={this.target} href={this.internalPrivacyPolicy}>Privacy policy</a>
+                  }
                 </div>
               </div>
-              <div class='sidebar__footer-wrapper-bottom-links'>
-                <a href='https://www.infineon.com/'>© 1999 - 2023 Infineon Technologies AG</a>
-              </div>
+              
+              {
+                this.copyrightText &&
+                <div class='sidebar__footer-wrapper-bottom-links'>
+                    <span>{this.copyrightText}</span>
+                </div>
+              }
             </div>
           </div>
         }